Use explicit downcast for 64-bit targets
GetLastError() is a DWORD (unsigned long) which is 64-bits on 64-bit targets,
while ExitProcess() takes UINT (unsigned int) which is always 32-bits (for
Windows targets).

+++ paste/paste.cpp
@@ -111,7 +111,7 @@ void Write(const wchar_t *text, DWORD outputHandle = STD_OUTPUT_HANDLE, DWORD ch
                                                    nullptr);
         result = WriteFile(hOut, _utf8Buffer.Buffer, bytesConverted, &charsWritten, nullptr);
         if (charsWritten != utf8ByteCount) {
-            ExitProcess(GetLastError());
+            ExitProcess((UINT) GetLastError());
         }
     }
